feat: Add feature flag for case defendant documents

+++ server/routes/index.js
@@ -253,38 +253,40 @@ module.exports = function Index ({ authenticationMiddleware }) {
 
   router.post('/:courtCode/hearing/:hearingId/defendant/:defendantId/summary/publish-edited-note', defaults, catchErrors(autoSaveHearingNoteEditHandler))
 
-  router.get('/:courtCode/hearing/:hearingId/defendant/:defendantId/summary/files/:fileId/delete', defaults, catchErrors(async (req, res) => {
-    const { params: { hearingId, defendantId, fileId } } = req
-    const { data } = await caseFiles.delete(hearingId, defendantId, fileId)
-    res.json(data)
-  }))
+  if (settings.enableCaseDefendantDocuments) {
+    router.get('/:courtCode/hearing/:hearingId/defendant/:defendantId/summary/files/:fileId/delete', defaults, catchErrors(async (req, res) => {
+      const { params: { hearingId, defendantId, fileId } } = req
+      const { data } = await caseFiles.delete(hearingId, defendantId, fileId)
+      res.json(data)
+    }))
 
-  router.get('/:courtCode/hearing/:hearingId/defendant/:defendantId/summary/files/:fileId/raw', (req, res, next) => {
-    const { params: { hearingId, defendantId, fileId } } = req
-    caseFiles.getRaw(
-      req,
-      res,
-      next,
-      proxyRes => {
-        if (proxyRes.statusCode >= 400) {
-          throw new Error(`${proxyRes.statusCode} - unable to download file.`)
-        }
-      },
-      hearingId,
-      defendantId,
-      fileId
-    )
-  })
+    router.get('/:courtCode/hearing/:hearingId/defendant/:defendantId/summary/files/:fileId/raw', (req, res, next) => {
+      const { params: { hearingId, defendantId, fileId } } = req
+      caseFiles.getRaw(
+        req,
+        res,
+        next,
+        proxyRes => {
+          if (proxyRes.statusCode >= 400) {
+            throw new Error(`${proxyRes.statusCode} - unable to download file.`)
+          }
+        },
+        hearingId,
+        defendantId,
+        fileId
+      )
+    })
 
-  router.post('/:courtCode/hearing/:hearingId/defendant/:defendantId/summary/files', (req, res, next) => {
-    const { params: { hearingId, defendantId } } = req
-    const formatter = data => {
-      // date formatter
-      data.datetime = moment(data.datetime).format('D MMMM YYYY')
-      return data
-    }
-    caseFiles.post(req, res, next, formatter, hearingId, defendantId)
-  })
+    router.post('/:courtCode/hearing/:hearingId/defendant/:defendantId/summary/files', (req, res, next) => {
+      const { params: { hearingId, defendantId } } = req
+      const formatter = data => {
+        // date formatter
+        data.datetime = moment(data.datetime).format('D MMMM YYYY')
+        return data
+      }
+      caseFiles.post(req, res, next, formatter, hearingId, defendantId)
+    })
+  }
 
   router.get('/:courtCode/hearing/:hearingId/defendant/:defendantId/summary', defaults, catchErrors(async (req, res) => {
     // build outcome types list for select controls
@@ -333,7 +335,8 @@ module.exports = function Index ({ authenticationMiddleware }) {
     templateValues.params.hearingOutcomesEnabled = hearingOutcomesEnabled
     templateValues.features = {
       hearingNotes: featuresToggles.hearingNotes.isEnabled(context),
-      hearingOutcomesEnabled
+      hearingOutcomesEnabled,
+      caseDefendantDocuments: settings.enableCaseDefendantDocuments
     }
     templateValues.outcomeTypes = outcomeTypes
     session.confirmedMatch = undefined
